UI/requirements.txt
flask
zeroconf
psutil
pillow
pypdf
docx2pdf
cryptography
imageio-ffmpeg
comtypes
python-pptx
reportlab
